API/requirements.txt
Django==3.0.5
Djangorestframework>=3.11.0,<3.12.0
gunicorn==19.7.1
django_neomodel==0.0.4
passlib==1.7.2
bcrypt==3.1.7
setuptools==46.1.3
PyJWT==1.7.1
cryptography==2.9
django-cors-headers==3.2.1
